refactor(loki): split alloy config into config toolset; address Ben's review
- Fix bug where RequiredLabels not in ApprovedLabels were injected with an
  "unknown" placeholder then immediately dropped by stage.label_keep.
  Union RequiredLabels into the kept set so soft-enforcement works.
- Move suggest_loki_alloy_label_config out of AddLokiTools into a new
  "config" toolset (tools/config.go, AddConfigTools), wired into main.go
  with a --disable-config flag. analyze_loki_labels stays under Loki.
- Refresh the stale top-of-file comment in tools/loki_label_analyzer.go
  to reflect the post-consolidation surface.
- Update README category from Loki to Config and add regression tests
  for the union behaviour and duplicate dedup.
