+++ README.md
@@ -311,7 +311,7 @@ Scopes define the specific resources that permissions apply to. Each action requ
 | `query_loki_stats`                | Loki        | Get statistics about log streams                                    | `datasources:query`                     | `datasources:uid:loki-uid`                          |
 | `query_loki_patterns`             | Loki        | Query detected log patterns to identify common structures           | `datasources:query`                     | `datasources:uid:loki-uid`                          |
 | `analyze_loki_labels`             | Loki        | Audit a Loki label strategy (live or static) and optionally diagnose query performance | `datasources:query`                     | `datasources:uid:loki-uid`                          |
-| `suggest_loki_alloy_label_config` | Loki        | Generate an Alloy `loki.process` snippet enforcing approved labels  | N/A                                     | N/A                                                 |
+| `suggest_loki_alloy_label_config` | Config      | Generate an Alloy `loki.process` snippet enforcing approved labels  | N/A                                     | N/A                                                 |
 | `query_influxdb`                  | InfluxDB    | Query InfluxDB using InfluxQL (v1) or Flux (v2)                     | `datasources:query`                     | `datasources:uid:influxdb-uid`                      |
 | `list_clickhouse_tables`          | ClickHouse* | List tables in a ClickHouse database                                | `datasources:query`                     | `datasources:uid:*`                                 |
 | `describe_clickhouse_table`       | ClickHouse* | Get table schema with column types                                  | `datasources:query`                     | `datasources:uid:*`                                 |
