+++ cmd/mcp-grafana/main.go
@@ -72,6 +72,7 @@ var categoryDescription = map[string]string{
 	"graphite":      "Graphite: Query Graphite datasources for metrics.",
 	"athena":        "Athena: Query Amazon Athena datasources via Grafana with SQL, macro substitution, and schema discovery.",
 	"api":           "API: Make authenticated HTTP requests to any Grafana API endpoint with optional jq-style response filtering.",
+	"config":        "Config: Generate operator-facing configuration snippets (e.g. Alloy label-enforcement pipelines).",
 }
 
 // disabledTools indicates whether each category of tools should be disabled.
@@ -83,7 +84,7 @@ type disabledTools struct {
 	dashboard, folder, oncall, asserts, sift, admin,
 	pyroscope, navigation, proxied, annotations, rendering, cloudwatch, write,
 	examples, clickhouse, snowflake, graphite,
-	runpanelquery, athena, plugin, api bool
+	runpanelquery, athena, plugin, api, config bool
 }
 
 // Configuration for the Grafana client.
@@ -102,7 +103,7 @@ type grafanaConfig struct {
 }
 
 func (dt *disabledTools) addFlags() {
-	flag.StringVar(&dt.enabledTools, "enabled-tools", "search,datasource,incident,prometheus,loki,alerting,dashboard,folder,oncall,asserts,sift,pyroscope,navigation,proxied,annotations,rendering,plugin,api", "A comma separated list of tools enabled for this server. Can be overwritten entirely or by disabling specific components, e.g. --disable-search.")
+	flag.StringVar(&dt.enabledTools, "enabled-tools", "search,datasource,incident,prometheus,loki,alerting,dashboard,folder,oncall,asserts,sift,pyroscope,navigation,proxied,annotations,rendering,plugin,api,config", "A comma separated list of tools enabled for this server. Can be overwritten entirely or by disabling specific components, e.g. --disable-search.")
 	flag.BoolVar(&dt.search, "disable-search", false, "Disable search tools")
 	flag.BoolVar(&dt.datasource, "disable-datasource", false, "Disable datasource tools")
 	flag.BoolVar(&dt.incident, "disable-incident", false, "Disable incident tools")
@@ -132,6 +133,7 @@ func (dt *disabledTools) addFlags() {
 	flag.BoolVar(&dt.athena, "disable-athena", false, "Disable Athena tools")
 	flag.BoolVar(&dt.plugin, "disable-plugin", false, "Disable plugin tools")
 	flag.BoolVar(&dt.api, "disable-api", false, "Disable API tools")
+	flag.BoolVar(&dt.config, "disable-config", false, "Disable config-generation tools")
 }
 
 func (gc *grafanaConfig) addFlags() {
@@ -187,6 +189,7 @@ func (dt *disabledTools) toolEntries() []toolEntry {
 		{tools.AddAthenaTools, dt.athena, "athena"},
 		{func(mcp *server.MCPServer) { tools.AddPluginTools(mcp, enableWriteTools) }, dt.plugin, "plugin"},
 		{func(mcp *server.MCPServer) { tools.AddAPITools(mcp, enableWriteTools) }, dt.api, "api"},
+		{tools.AddConfigTools, dt.config, "config"},
 	}
 }
