+++ tools/loki.go
@@ -915,7 +915,9 @@ var QueryLokiPatterns = mcpgrafana.MustTool(
 	mcp.WithReadOnlyHintAnnotation(true),
 )
 
-// AddLokiTools registers all Loki tools with the MCP server
+// AddLokiTools registers all Loki tools with the MCP server.
+// Config-generation tools (e.g. suggest_loki_alloy_label_config) live in
+// the separate "config" category — see tools.AddConfigTools.
 func AddLokiTools(mcp *server.MCPServer) {
 	ListLokiLabelNames.Register(mcp)
 	ListLokiLabelValues.Register(mcp)
