+++ tools/loki_label_analyzer.go
@@ -15,13 +15,12 @@ import (
 // The functions in this file translate the Grafana Professional Services
 // Loki label strategy into deterministic, machine-readable tools:
 //
-//   - get_loki_label_cardinality          (live)
-//   - audit_loki_label_strategy           (live + static)
-//   - diagnose_loki_query_performance     (live + static)
-//   - suggest_loki_alloy_label_config     (static)
+//   - analyze_loki_labels  (live + static; optionally diagnoses query perf)
 //
 // The scoring rules live in this file so the verdicts are easy to audit
-// and adjust without touching transport code.
+// and adjust without touching transport code. The Alloy config-generation
+// tool (suggest_loki_alloy_label_config) is a separate toolset; see
+// tools/config.go.
 
 // ---------------------------------------------------------------------------
 // Cardinality bands
@@ -922,100 +921,6 @@ func buildPerfSummary(findings []QueryPerfFinding, stats *Stats) string {
 }
 
 // ---------------------------------------------------------------------------
-// suggest_loki_alloy_label_config
-// ---------------------------------------------------------------------------
-
-// SuggestLokiAlloyLabelConfigParams describes the desired Alloy pipeline.
-type SuggestLokiAlloyLabelConfigParams struct {
-	ApprovedLabels    []string `json:"approvedLabels" jsonschema:"required,description=Labels to keep on the index."`
-	RequiredLabels    []string `json:"requiredLabels,omitempty" jsonschema:"description=Labels that get an 'unknown' placeholder when missing."`
-	NormalizeLogLevel bool     `json:"normalizeLogLevel,omitempty"`
-	ComponentName     string   `json:"componentName,omitempty"`
-	ForwardTo         string   `json:"forwardTo,omitempty"`
-}
-
-// SuggestLokiAlloyLabelConfigResult is the rendered snippet plus notes.
-type SuggestLokiAlloyLabelConfigResult struct {
-	Alloy string   `json:"alloy"`
-	Notes []string `json:"notes,omitempty"`
-}
-
-func suggestLokiAlloyLabelConfig(_ context.Context, args SuggestLokiAlloyLabelConfigParams) (*SuggestLokiAlloyLabelConfigResult, error) {
-	if len(args.ApprovedLabels) == 0 {
-		return nil, fmt.Errorf("approvedLabels must contain at least one label")
-	}
-
-	componentName := args.ComponentName
-	if componentName == "" {
-		componentName = "enforce_labels"
-	}
-	forwardTo := args.ForwardTo
-	if forwardTo == "" {
-		forwardTo = "loki.write.default.receiver"
-	}
-
-	approved := make([]string, len(args.ApprovedLabels))
-	copy(approved, args.ApprovedLabels)
-	sort.Strings(approved)
-
-	var b strings.Builder
-	fmt.Fprintf(&b, "loki.process %q {\n", componentName)
-	fmt.Fprintf(&b, "  forward_to = [%s]\n\n", forwardTo)
-
-	if args.NormalizeLogLevel {
-		b.WriteString("  // Normalize log level: I/Info/INFO -> info, etc. Patterns are anchored so partial matches\n")
-		b.WriteString("  // inside other level values (e.g. \"CRITICAL\", \"trace\") aren't mangled.\n")
-		b.WriteString("  stage.replace {\n    source = \"level\"\n    expression = \"^(?i)I(nfo)?$\"\n    replace = \"info\"\n  }\n")
-		b.WriteString("  stage.replace {\n    source = \"level\"\n    expression = \"^(?i)W(arn(ing)?)?$\"\n    replace = \"warn\"\n  }\n")
-		b.WriteString("  stage.replace {\n    source = \"level\"\n    expression = \"^(?i)E(rr(or)?)?$\"\n    replace = \"error\"\n  }\n")
-		b.WriteString("  stage.replace {\n    source = \"level\"\n    expression = \"^(?i)D(ebug)?$\"\n    replace = \"debug\"\n  }\n")
-		b.WriteString("  stage.labels {\n    values = { level = \"\" }\n  }\n\n")
-	}
-
-	for _, name := range args.RequiredLabels {
-		fmt.Fprintf(&b, "  // Soft enforcement: inject \"unknown\" when %q is missing so violations are queryable.\n", name)
-		fmt.Fprintf(&b, "  stage.template {\n    source = %q\n    template = \"{{ if .Value }}{{ .Value }}{{ else }}unknown{{ end }}\"\n  }\n", name)
-		fmt.Fprintf(&b, "  stage.labels {\n    values = { %s = \"\" }\n  }\n\n", name)
-	}
-
-	b.WriteString("  // Final enforcement: keep only the approved label set.\n")
-	b.WriteString("  stage.label_keep {\n    values = [")
-	for i, label := range approved {
-		if i > 0 {
-			b.WriteString(", ")
-		}
-		fmt.Fprintf(&b, "%q", label)
-	}
-	b.WriteString("]\n  }\n")
-	b.WriteString("}\n")
-
-	notes := []string{
-		"stage.label_keep is the last enforcement step — anything not listed in approvedLabels is dropped.",
-	}
-	if !args.NormalizeLogLevel {
-		notes = append(notes, "Consider enabling normalizeLogLevel if level values arrive in mixed casing.")
-	}
-	if len(args.RequiredLabels) > 0 {
-		notes = append(notes, "Soft enforcement is a runway to hard enforcement; once `unknown` counts trend to zero, switch to dropping signals with missing required labels.")
-	}
-
-	return &SuggestLokiAlloyLabelConfigResult{
-		Alloy: b.String(),
-		Notes: notes,
-	}, nil
-}
-
-// SuggestLokiAlloyLabelConfig is the registered tool wrapper.
-var SuggestLokiAlloyLabelConfig = mcpgrafana.MustTool(
-	"suggest_loki_alloy_label_config",
-	"Generates an Alloy loki.process snippet enforcing an approved label set via stage.label_keep, with optional log-level normalisation and soft-enforcement placeholders.",
-	suggestLokiAlloyLabelConfig,
-	mcp.WithTitleAnnotation("Suggest Alloy label enforcement config"),
-	mcp.WithIdempotentHintAnnotation(true),
-	mcp.WithReadOnlyHintAnnotation(true),
-)
-
-// ---------------------------------------------------------------------------
 // Registration helper
 // ---------------------------------------------------------------------------
 
@@ -1023,7 +928,6 @@ var SuggestLokiAlloyLabelConfig = mcpgrafana.MustTool(
 // AddLokiTools calls this so the standard Loki bundle includes them.
 func AddLokiLabelAnalyzerTools(s *server.MCPServer) {
 	AnalyzeLokiLabels.Register(s)
-	SuggestLokiAlloyLabelConfig.Register(s)
 }
 
 // ---------------------------------------------------------------------------
